+++ .gitignore
@@ -45,3 +45,6 @@ yarn-error.log*
 *.tsbuildinfo
 next-env.d.ts
 
+# Serwist
+public/sw*
+public/swe-worker*
