+++ package.json
@@ -35,6 +35,9 @@
         "@entur/typography": "^1.8.39",
         "@graphql-codegen/cli": "^5.0.2",
         "@internationalized/date": "^3.5.4",
+        "@serwist/next": "^9.0.7",
+        "@serwist/precaching": "^9.0.7",
+        "@serwist/sw": "^9.0.7",
         "chart.js": "^4.4.3",
         "date-fns": "^3.6.0",
         "framer-motion": "^11.2.12",
@@ -44,7 +47,6 @@
         "next": "^14.2.4",
         "next-auth": "^4.24.7",
         "next-logger": "^5.0.0",
-        "next-pwa": "^5.6.0",
         "pino": "^9.3.2",
         "react": "^18.3.1",
         "react-chartjs-2": "^5.2.0",
@@ -71,6 +73,7 @@
         "postcss": "^8",
         "prettier": "3.3.2",
         "raw-loader": "^4.0.2",
+        "serwist": "^9.0.7",
         "tailwindcss": "^3.4.1",
         "typescript": "^5"
     },
