+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
     "compilerOptions": {
-        "lib": ["dom", "dom.iterable", "esnext"],
+        "lib": ["dom", "dom.iterable", "esnext", "webworker"],
         "allowJs": true,
         "skipLibCheck": true,
         "strict": true,
@@ -12,6 +12,7 @@
         "isolatedModules": true,
         "jsx": "preserve",
         "incremental": true,
+        "types": ["@serwist/next/typings"],
         "plugins": [
             {
                 "name": "next"
@@ -28,5 +29,5 @@
         ".next/types/**/*.ts",
         "middleware.ts"
     ],
-    "exclude": ["node_modules"]
+    "exclude": ["node_modules", "public/sw.js"]
 }
